// packages/runtime-vapor/js-reactive/types.d.ts
export type ObjectLike = Record<PropertyKey, unknown>
export interface HookHost extends ObjectLike {
  __hooks?: unknown
}
export interface HookContainer extends ObjectLike {
  states: unknown[]
  index: number
  __forcedIndex?: number
  __idMap?: Map<unknown, unknown>
}
export interface HookFrame {
  instance: unknown
  hooks?: HookContainer
}
export interface HookCarrier {
  getCurrentInstance(): unknown
  renderHooks<T>(instance: unknown, render: () => T): T
}
export interface HookContext extends HookCarrier {
  setCurrentInstance(instance: unknown): void
  vaporWithHookId<T>(id: unknown, runner: () => T): T
  withHookSlot<T>(factory: () => T): T
}